add indicator for Safe logged in user on the header
